smart_home/requirements.txt
Django>=4.2
djangorestframework
requests
